Move MockPncGateway to tests/helpers folder
