+++ packages/core/comparison/lib/comparePhase1.ts
@@ -13,10 +13,10 @@ import { extractExceptionsFromXml, parseAhoXml } from "../../lib/parse/parseAhoX
 import serialiseToXml from "../../lib/serialise/ahoXml/serialiseToXml"
 import getMessageType from "../../phase1/lib/getMessageType"
 import phase1Handler from "../../phase1/phase1"
+import MockPncGateway from "../../tests/helpers/MockPncGateway"
 import generateMockPncQueryResultFromAho from "./generateMockPncQueryResultFromAho"
 import getPncQueryTimeFromAho from "./getPncQueryTimeFromAho"
 import isIntentionalDifference from "./isIntentionalDifference"
-import MockPncGateway from "./MockPncGateway"
 import parseIncomingMessage from "./parseIncomingMessage"
 import { sortExceptions } from "./sortExceptions"
 import { sortTriggers } from "./sortTriggers"
