+++ packages/core/comparison/lib/comparePhase3.ts
@@ -19,9 +19,9 @@ import { isPncLockError } from "../../phase3/exceptions/generatePncUpdateExcepti
 import { MAXIMUM_PNC_LOCK_ERROR_RETRIES } from "../../phase3/lib/updatePnc"
 import phase3Handler from "../../phase3/phase3"
 import getPncOperationsFromPncUpdateDataset from "../../phase3/tests/helpers/getPncOperationsFromPncUpdateDataset"
+import MockPncGateway from "../../tests/helpers/MockPncGateway"
 import { isPncUpdateDataset } from "../../types/PncUpdateDataset"
 import extractAuditLogEventCodes from "./extractAuditLogEventCodes"
-import MockPncGateway from "./MockPncGateway"
 import parseIncomingMessage from "./parseIncomingMessage"
 import { sortExceptions } from "./sortExceptions"
 import { sortTriggers } from "./sortTriggers"
