+++ packages/core/comparison/lib/comparePncMatching.ts
@@ -8,9 +8,9 @@ import type PncComparisonResultDetail from "../types/PncComparisonResultDetail"
 import CoreAuditLogger from "../../lib/auditLog/CoreAuditLogger"
 import { parseAhoXml } from "../../lib/parse/parseAhoXml"
 import CorePhase1 from "../../phase1/phase1"
+import MockPncGateway from "../../tests/helpers/MockPncGateway"
 import generateMockPncQueryResultFromAho from "./generateMockPncQueryResultFromAho"
 import getPncQueryTimeFromAho from "./getPncQueryTimeFromAho"
-import MockPncGateway from "./MockPncGateway"
 import parseIncomingMessage from "./parseIncomingMessage"
 import summariseMatching from "./summariseMatching"
