+++ packages/core/phase1/enrichAho/enrichFunctions/enrichWithPncQuery.test.ts
@@ -8,12 +8,12 @@ import type AuditLogger from "../../../types/AuditLogger"
 import type PncGatewayInterface from "../../../types/PncGatewayInterface"
 import type { PncQueryResult } from "../../../types/PncQueryResult"
 
-import MockPncGateway from "../../../comparison/lib/MockPncGateway"
 import CoreAuditLogger from "../../../lib/auditLog/CoreAuditLogger"
 import errorPaths from "../../../lib/exceptions/errorPaths"
 import parseSpiResult from "../../../lib/parse/parseSpiResult"
 import transformSpiToAho from "../../../lib/parse/transformSpiToAho"
 import { PncApiError } from "../../../lib/pnc/PncGateway"
+import MockPncGateway from "../../../tests/helpers/MockPncGateway"
 import generateMessage from "../../tests/helpers/generateMessage"
 import generateMockPncQueryResult from "../../tests/helpers/generateMockPncQueryResult"
 import enrichWithPncQuery from "./enrichWithPncQuery"
