+++ packages/core/phase1/scripts/compareResults.ts
@@ -8,10 +8,10 @@ import type Exception from "../../types/Exception"
 import type { Trigger } from "../../types/Trigger"
 import type Phase1Result from "../types/Phase1Result"
 
-import MockPncGateway from "../../comparison/lib/MockPncGateway"
 import CoreAuditLogger from "../../lib/auditLog/CoreAuditLogger"
 import parseSpiResult from "../../lib/parse/parseSpiResult"
 import transformSpiToAho from "../../lib/parse/transformSpiToAho"
+import MockPncGateway from "../../tests/helpers/MockPncGateway"
 import CorePhase1 from "../phase1"
 import generateMockPncQueryResult from "../tests/helpers/generateMockPncQueryResult"
 import { Phase1ResultType } from "../types/Phase1Result"
