+++ packages/core/phase3/phase3.comparison.test.ts
@@ -11,7 +11,6 @@ import type { PncException } from "../types/Exception"
 import type Phase3Result from "./types/Phase3Result"
 
 import { normalisePncOperations } from "../comparison/lib/comparePhase3"
-import MockPncGateway from "../comparison/lib/MockPncGateway"
 import parseIncomingMessage from "../comparison/lib/parseIncomingMessage"
 import CoreAuditLogger from "../lib/auditLog/CoreAuditLogger"
 import saveErrorListRecord from "../lib/database/saveErrorListRecord"
@@ -27,6 +26,7 @@ import {
   sql
 } from "../tests/helpers/comparison/e2eComparisonTestsHelpers"
 import getComparisonTests from "../tests/helpers/comparison/getComparisonTests"
+import MockPncGateway from "../tests/helpers/MockPncGateway"
 import { isPncUpdateDataset } from "../types/PncUpdateDataset"
 import { isPncLockError } from "./exceptions/generatePncUpdateExceptionFromMessage"
 import { MAXIMUM_PNC_LOCK_ERROR_RETRIES } from "./lib/updatePnc"
