+++ packages/core/phase3/phase3.test.ts
@@ -5,9 +5,9 @@ import { isError } from "@moj-bichard7/common/types/Result"
 import type Phase3Result from "./types/Phase3Result"
 import type PncUpdateRequestError from "./types/PncUpdateRequestError"
 
-import MockPncGateway from "../comparison/lib/MockPncGateway"
 import CoreAuditLogger from "../lib/auditLog/CoreAuditLogger"
 import { PncApiError } from "../lib/pnc/PncGateway"
+import MockPncGateway from "../tests/helpers/MockPncGateway"
 import { PncOperation } from "../types/PncOperation"
 import phase3 from "./phase3"
 import generatePncUpdateDatasetWithOperations from "./tests/helpers/generatePncUpdateDatasetWithOperations"
