+++ packages/core/scripts/testServer.ts
@@ -5,11 +5,11 @@ import express from "express"
 
 import type { PncQueryResult } from "../types/PncQueryResult"
 
-import MockPncGateway from "../comparison/lib/MockPncGateway"
 import parseIncomingMessage from "../comparison/lib/parseIncomingMessage"
 import CoreAuditLogger from "../lib/auditLog/CoreAuditLogger"
 import CorePhase1 from "../phase1/phase1"
 import CorePhase2 from "../phase2/phase2"
+import MockPncGateway from "../tests/helpers/MockPncGateway"
 import Phase from "../types/Phase"
 
 const app = express()
