+++ packages/core/tsconfig.build.json
@@ -7,6 +7,7 @@
     "phase2/tests/helpers/PartialPncUpdateDataset.ts",
     "phase2/tests/fixtures/helpers/generateFakePncUpdateDataset.ts",
     "phase3/tests/helpers/getPncErrorMessages.ts",
-    "phase3/tests/helpers/getPncOperationsFromPncUpdateDataset.ts"
+    "phase3/tests/helpers/getPncOperationsFromPncUpdateDataset.ts",
+    "tests/helpers/MockPncGateway.ts"
   ]
 }
